@@ -2,16 +2,16 @@ package fr.ippon.search.functional.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Data
+@Getter
 @Builder
-@AllArgsConstructor
 @NoArgsConstructor
+@AllArgsConstructor
 public class Company {
     private String name;
     private String streetName;
     private String country;
-    private String status;
+    private Status status;
 }
