@@ -7,7 +7,6 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -18,13 +17,11 @@ public class SearchService {
     private final ProviderClient providerClient;
 
     public List<Company> searchCompanies(Criteria criteria) {
-        // On appelle ici notre ProviderApi
         try {
             return providerClient.searchCompanies(criteria);
         } catch (Exception exception) {
-            // On catch ici une Exception dans le cas où l’appel n’a pas renvoyé un status code 200 voire 404
             log.warn("Call to provider api went wrong", exception);
-            return new ArrayList<>();
+            return List.of();
         }
     }
 
